requirements.txt
flask==3.1.0
flask-cors==5.0.0
python-dotenv==0.21.0
fauna==2.4.0
